+++ package.json
@@ -13,6 +13,7 @@
     "express": "^4.18.2"
   },
   "devDependencies": {
-    "jest": "^29.7.0"
+    "jest": "^29.7.0",
+    "supertest": "^6.3.4"
   }
 }
