+++ src/index.js
@@ -1,62 +1,6 @@
-const express = require('express');
-const bodyParser = require('body-parser');
+const app = import('./api')
 
-const app = express();
 const port = 3000;
-
-//ola jjjj
-app.use(bodyParser.json());
-
-let tasks = [
-  { id: 1, title: 'Task 1', description: 'Do something' },
-  { id: 2, title: 'Task 2', description: 'Do something else' },
-];
-
-// Get all tasks
-app.get('/tasks', (req, res) => {
-  res.json(tasks);
-});
-
-// Get a specific task
-app.get('/tasks/:id', (req, res) => {
-  const taskId = parseInt(req.params.id);
-  const task = tasks.find((t) => t.id === taskId);
-
-  if (task) {
-    res.json(task);
-  } else {
-    res.status(404).json({ error: 'Task not found' });
-  }
-});
-
-// Create a new task
-app.post('/tasks', (req, res) => {
-  const newTask = req.body;
-  newTask.id = tasks.length + 1;
-  tasks.push(newTask);
-  res.status(201).json(newTask);
-});
-
-// Update a task
-app.put('/tasks/:id', (req, res) => {
-  const taskId = parseInt(req.params.id);
-  const updatedTask = req.body;
-  const index = tasks.findIndex((t) => t.id === taskId);
-
-  if (index !== -1) {
-    tasks[index] = { ...tasks[index], ...updatedTask };
-    res.json(tasks[index]);
-  } else {
-    res.status(404).json({ error: 'Task not found' });
-  }
-});
-
-// Delete a task
-app.delete('/tasks/:id', (req, res) => {
-  const taskId = parseInt(req.params.id);
-  tasks = tasks.filter((t) => t.id !== taskId);
-  res.sendStatus(204);
-});
 
 app.listen(port, () => {
   console.log(`Server is running on http://localhost:${port}`);
